itinerary refactored
