+++ client/src/helpers/fetchUtils.js
@@ -1,0 +1,4 @@
+const fetchJson = (url) =>
+  fetch(url).then(res => res.json());
+
+export { fetchJson };
